TIP-221: Move CompletenessManager
